Fix additional ruff check violations
  1. PD003 (.isnull → .isna): Fixed all 4 instances
    - Updated deprecated pandas methods to modern equivalents
  2. PD011 (.values → .to_numpy()): Fixed many instances
    - Converted pandas .values to more performant .to_numpy()
    - Fixed dozens of instances across the codebase
  3. RUF005 (List concatenation → unpacking): Fixed all 13 instances
    - Modernized list concatenation using unpacking ([*list1, *list2])
    - Updated subprocess calls, column definitions, and tuple constructions
  4. RUF046 (Unnecessary int casting): Fixed all 4 instances
    - Removed redundant int(round(...)) since round() returns int in Python 3
    - Fixed int(math.ceil(...)) since math.ceil() returns int in Python 3
  5. TC003 (Type-checking imports): Fixed 1 instance
    - Moved collections.abc.Iterable import to TYPE_CHECKING block

+++ cnvlib/smoothing.py
@@ -36,7 +36,7 @@ def check_inputs(x, width, as_series=True, weights=None):
 def _width2wing(width, x, min_wing=3):
     """Convert a fractional or absolute width to integer half-width ("wing")."""
     if 0 < width < 1:
-        wing = int(math.ceil(len(x) * width * 0.5))
+        wing = math.ceil(len(x) * width * 0.5)
     elif width >= 2 and int(width) == width:
         # Ensure window width <= len(x) to avoid TypeError
         width = min(width, len(x) - 1)
@@ -125,7 +125,7 @@ def guess_window_size(x, weights=None):
     else:
         sd = descriptives.weighted_std(x, weights)
     width = 4 * sd * len(x) ** (4 / 5)
-    width = max(3, int(round(width)))
+    width = max(3, round(width))
     width = min(len(x), width)
     return width
